AIML_Assignment/requirements.txt
fastapi
uvicorn
sqlmodel
aiosqlite
pydantic
sqlalchemy
ultralytics
opencv-python-headless
python-multipart
chromadb
streamlit
requests
# torch --index-url https://download.pytorch.org/whl/cpu # Optional: For smaller CPU-only build


easyocr

supabase
python-dotenv
